Improve 32 bit support

+++ 3D/src/common/types.h
@@ -1,3 +1,3 @@
 #pragma once
 
-typedef size_t gSize_t;
+typedef unsigned int gSize_t;
